// FoodDiary.Web.Client/src/app/shared/api/sdk/generated/model/client-task-http-response.ts
/**
 * FoodDiary User API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface ClientTaskHttpResponse {
    clientUserId?: string;
    createdAtUtc?: string;
    details?: string | null;
    dietologistUserId?: string;
    dueAtUtc?: string | null;
    id?: string;
    isOverdue?: boolean;
    status?: string;
    statusChangedAtUtc?: string | null;
    title?: string;
}