// FoodDiary.Web.Client/src/app/shared/api/sdk/generated/model/recommendation-comment-http-response.ts
/**
 * FoodDiary User API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface RecommendationCommentHttpResponse {
    authorEmail?: string | null;
    authorFirstName?: string | null;
    authorLastName?: string | null;
    authorUserId?: string;
    createdAtUtc?: string;
    id?: string;
    recommendationId?: string;
    text?: string;
}