// FoodDiary.Web.Client/src/app/shared/api/sdk/generated/model/recommendation-http-response.ts
/**
 * FoodDiary User API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface RecommendationHttpResponse {
    createdAtUtc?: string;
    dietologistFirstName?: string | null;
    dietologistLastName?: string | null;
    dietologistUserId?: string;
    id?: string;
    isRead?: boolean;
    readAtUtc?: string | null;
    text?: string;
}